Перенес миксин PrevNextMixin в mixins
